% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careroll.R
\name{careroll}
\alias{careroll}
\title{Download Medicare Beneficiary Enrollment Data}
\usage{
careroll()
}
\value{
A \code{tibble}
}
\description{
Download Medicare Beneficiary Enrollment Data
}
\examples{
\dontrun{
careroll()
}
}
