% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careroll.R
\name{change_year}
\alias{change_year}
\title{Download and sort Medicare Beneficiary Enrollment Data}
\usage{
change_year(df, col)
}
\arguments{
\item{df}{year, month}

\item{col}{national, state, county}
}
\value{
A \code{tibble}
}
\description{
Download and sort Medicare Beneficiary Enrollment Data
}
\examples{
\dontrun{
change_year()
}
}
