% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dict.R
\docType{data}
\name{data_dict}
\alias{data_dict}
\title{Data Dictionary}
\format{
\subsection{\code{data_dict}}{

A tibble with 15 rows and 2 columns:
\describe{
\item{var}{variable name}
\item{desc}{variable description}
}
}
}
\usage{
data_dict
}
\description{
Data Dictionary
}
\keyword{datasets}
