% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careroll.R
\name{levels}
\alias{levels}
\title{Download and sort Medicare Beneficiary Enrollment Data}
\usage{
levels(
  period = c("year", "month"),
  level = c("national", "state", "county"),
  group = c("total", "aged", "disabled", "partAB", "partD")
)
}
\arguments{
\item{period}{year, month}

\item{level}{national, state, county}

\item{group}{total, aged, disabled, partAB, partD}
}
\value{
A \code{tibble}
}
\description{
Download and sort Medicare Beneficiary Enrollment Data
}
\examples{
\dontrun{
levels()
}
}
